✅ feat(validate): Validate the commit msg in .git is from this cli
Prior to this change, there was no way to enforce users to commit only
through this cli tool

This change adds a --validate-commit flag to the "commit" command, that
throws an error if the current COMMIT_EDITMSG in .git dir does not match
the last message created via the "commit" command.

This is expected to be used on the "commit-msg" git hook. This commit
does not add this validation to the git hook. That still has to be done
manually.

- Original Branch: master