✅ feat(validate): Ensure global and repo level settings exist
Prior to this change, there was no way of telling whether a user is using
this cli when making commits. So if the commit validation was being used
and someone without this cli tried to commit, it would always fail.

This change asks the user to supply settings to decide whether we should
force validation and whether to skip errors that occur during validation.

- Original Branch: master