✅ feat(validate): Let the "-u" flag only validate based of user settings
Prior to this change, if the validate command was added to a repo where
some people do not intend to use the cli, then it would fail for them all
the time.

This change allows you to pass the "-u" flag to "--validate-commit" which
will only validate the commit by looking at the users settings and seeing
then whether it should validate. This means anyone without this cli can
commit how they want, but those with this cli can still enforce using it.

- Original Branch: master